refactor: 將getTopFollowedUsers移到middleware
