+++ controllers/followship-controller.js
@@ -48,26 +48,6 @@ const followshipController = {
       })
       .then(() => res.redirect('back'))
       .catch(err => next(err))
-  },
-  getTopFollowedUsers: (req, res, next) => {
-    const loginUser = helpers.getUser(req)
-    const topFollowedUsersNumber = 10
-    return User.findAll({
-      include: [{ model: User, as: 'Followers' }],
-      where: { role: 'user', id: { [Op.not]: loginUser.id } }
-    })
-      .then(users => {
-        const result = users
-          .map(user => ({
-            ...user.toJSON(),
-            followerCount: user.Followers.length,
-            isFollowed: loginUser.Followings?.some(f => f.id === user.id)
-          }))
-          .sort((a, b) => b.followerCount - a.followerCount)
-        req.followingData = result.slice(0, topFollowedUsersNumber - 1)
-        return next()
-      })
-      .catch(err => next(err))
   }
 }
